requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==4.2.20
django-filter==25.1
djangorestframework==3.15.2
drf-yasg==1.21.10
exceptiongroup==1.2.2
flake8==7.1.2
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.1.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.5
pytest-django==4.10.0
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
